# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Valora MCP Server API
  description: |
    # Valora MCP Server API Documentation
    
    The Valora MCP Server provides a secure, scalable Memory Container Protocol implementation
    with enterprise-grade features including API key management, integrations, and chat continuity.
    
    ## Features
    
    - **Memory Management**: Store, retrieve, and search memories
    - **Chat Continuity**: Import and export conversations for AI agents
    - **Integration System**: Webhooks and plugins for external services
    - **Security**: API key authentication with permissions and rate limiting
    - **Monitoring**: Health checks and usage statistics
    
    ## Authentication
    
    All API endpoints require authentication using API keys. Include your API key in the
    Authorization header:
    
    ```
    Authorization: Bearer your-api-key
    ```
    
    ## Rate Limiting
    
    API requests are rate-limited per API key:
    - Default: 100 requests per minute, 1000 per hour
    - Configurable per API key
    - Rate limit headers included in responses
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns detailed error messages:
    - `400` - Bad Request (invalid input)
    - `401` - Unauthorized (invalid API key)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found (resource doesn't exist)
    - `429` - Too Many Requests (rate limit exceeded)
    - `500` - Internal Server Error
    
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /integrations/status:
    get:
      summary: Get Integration Status
      description: Get status of all integrations (webhooks and plugins)
      tags:
        - Integrations
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Integration status
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 5
                      enabled:
                        type: integer
                        example: 3
                      disabled:
                        type: integer
                        example: 2
                  plugins:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 2
                      enabled:
                        type: integer
                        example: 1
                      disabled:
                        type: integer
                        example: 1
                  events:
                    type: array
                    items:
                      type: string
                    example:
                      - "memory.created"
                      - "memory.updated"
                      - "memory.deleted"
                      - "chat.imported"
                      - "search.performed"
                      - "export.completed"
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /integrations/webhooks:
    get:
      summary: List Webhooks
      description: Get all registered webhooks
      tags:
        - Integrations
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Webhook'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Register Webhook
      description: Register a new webhook for event notifications
      tags:
        - Integrations
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - events
              properties:
                url:
                  type: string
                  format: uri
                  description: Webhook URL
                  example: "https://api.example.com/webhooks/valora"
                events:
                  type: array
                  items:
                    type: string
                    enum:
                      - "memory.created"
                      - "memory.updated"
                      - "memory.deleted"
                      - "chat.imported"
                      - "search.performed"
                      - "export.completed"
                  description: Events to subscribe to
                  example: ["memory.created", "chat.imported"]
                headers:
                  type: object
                  description: Additional headers to include
                  example:
                    "X-Custom-Header": "custom-value"
                retryPolicy:
                  type: object
                  properties:
                    maxRetries:
                      type: integer
                      default: 3
                    backoffMs:
                      type: integer
                      default: 1000
                    timeoutMs:
                      type: integer
                      default: 5000
                enabled:
                  type: boolean
                  default: true
      responses:
        '201':
          description: Webhook registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          description: Invalid webhook configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /integrations/webhooks/{id}:
    get:
      summary: Get Webhook
      description: Get details of a specific webhook
      tags:
        - Integrations
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Webhook ID
          schema:
            type: string
      responses:
        '200':
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update Webhook
      description: Update webhook configuration
      tags:
        - Integrations
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Webhook ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                headers:
                  type: object
                retryPolicy:
                  type: object
                enabled:
                  type: boolean
      responses:
        '200':
          description: Webhook updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete Webhook
      description: Delete a webhook
      tags:
        - Integrations
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Webhook ID
          schema:
            type: string
      responses:
        '200':
          description: Webhook deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Webhook deleted successfully"
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key for authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid API key"
        code:
          type: string
          description: Error code
          example: "INVALID_API_KEY"
        details:
          type: object
          description: Additional error details
...
    Webhook:
      type: object
      properties:
        id:
          type: string
          example: "webhook_123456789"
        url:
          type: string
          format: uri
          example: "https://api.example.com/webhooks/valora"
        events:
          type: array
          items:
            type: string
          example: ["memory.created", "chat.imported"]
        headers:
          type: object
          example:
            "X-Custom-Header": "custom-value"
        retryPolicy:
          type: object
          properties:
            maxRetries:
              type: integer
              example: 3
            backoffMs:
              type: integer
              example: 1000
            timeoutMs:
              type: integer
              example: 5000
        enabled:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...